リファクタリング
・バージョン照会処理クラスを外出し
